flaremo-tools/requirements.txt
Django==5.1.4
gunicorn==23.0.0
whitenoise==6.7.0
python-docx==1.1.2
Pillow==10.4.0
deep-translator==1.11.4
django-tinymce
cloudinary
requests
gradio_client 
moviepy